First commit, testing new branch. Package refactoring
